+++ tests/testthat/test-mltools.R
@@ -1,7 +1,7 @@
 cloud <- ichimoku(sample_ohlc_data)
-grid <- mlgrid(cloud, y = "logret", type = "numeric")
-grid2 <- mlgrid(cloud, y = "ret", type = "boolean", dir = "short", unique = FALSE)
-grid3 <- mlgrid(cloud, y = "none", type = "z-score", format = "matrix")
+grid <- mlgrid(cloud, y = "none", type = "numeric")
+grid2 <- mlgrid(cloud, y = "logret", type = "boolean", dir = "short", unique = FALSE)
+grid3 <- mlgrid(cloud, y = "ret", k = 3, type = "z-score", format = "matrix")
 stratlist <- autostrat(cloud, n = 2, dir = "short", quietly = TRUE)
 
 test_that("autostrat ok", {
@@ -13,12 +13,13 @@ test_that("autostrat ok", {
 
 test_that("mlgrid ok", {
   expect_s3_class(grid, "data.frame")
-  expect_identical(dim(grid), c(153L, 38L))
+  expect_identical(dim(grid), c(155L, 37L))
   expect_s3_class(grid2, "data.frame")
   expect_identical(dim(grid2), c(153L, 75L))
   expect_type(grid3, "double")
-  expect_identical(dim(grid3), c(155L, 37L))
+  expect_identical(dim(grid3), c(151L, 38L))
   expect_error(mlgrid(sample_ohlc_data), regexp = "ichimoku object")
+  expect_error(mlgrid(cloud, k = NA), regexp = "invalid value")
 })
 
 test_that("relative ok", {
@@ -32,7 +33,7 @@ test_that("relative ok", {
 test_that("look ok", {
   expect_length(expect_type(look(cloud), "list"), 3L)
   expect_length(expect_type(look(stratlist[[1L]]), "list"), 4L)
-  expect_length(expect_type(look(grid), "list"), 4L)
+  expect_length(expect_type(look(grid), "list"), 5L)
   expect_length(expect_type(look(stratlist), "list"), 2L)
   expect_s3_class(look(stratlist, which = 1), "ichimoku")
   expect_null(expect_invisible(look(sample_ohlc_data)))